Almost Finished With Project One
With it being rather short, I've already made quite a ways into almost finishing the first project of the stack, even with a rather sluggish pace of trying to figure out some things. To my surprised though, everything seems to actually be working all fine and good, so perhaps that did pay off in the end. I'm currently thinking of getting the whole thing done tonight, including the documentation as well.